Perform both combined and individual parses for multiple sources.
The original behavior is to only do the combined parse, where passing
multiple sources takes them all to be of the same parser type.

For BOMD, ORCA 4.0 is very different from < 4.0. In order to have
finer control over the print level, the SCF and coordinate information
is forcibly separated from the main output file, so they need a
combined parse. However, we don't currently have a built-in XYZ
reader, and the XYZ file is improperly recognized as an ORCA file and
can't be read from.

Current disadvantages:

* For this to work (and the unit test to pass), Open Babel is
required, otherwise the error "has no attribute 'atomcoords'" will
occur.
* This is handled by ccread, not ccopen.
* Does extra parsing work.